fix(calc): 피드백 개선
