--- src/main/java/calculator/Calculator.java
@@ -1,12 +1,0 @@
-package calculator;
-
-public class Calculator {
-    static int add(int i, int j) {
-        return i + j;
-    }
-
-    static int subtract(int i, int j) {
-        return i - j;
-    }
-
-}
